+++ lua/kubectl/views/deployments/views.lua
@@ -1,8 +1,8 @@
-local actions = require("kubectl.actions")
-local commands = require("kubectl.commands")
-local deployments = require("kubectl.deployments")
+local actions = require("kubectl.actions.actions")
+local commands = require("kubectl.actions.commands")
+local deployments = require("kubectl.views.deployments")
 local find = require("kubectl.utils.find")
-local tables = require("kubectl.view.tables")
+local tables = require("kubectl.utils.tables")
 
 local M = {}
