+++ lua/kubectl/views/events/init.lua
