+++ lua/kubectl/views/nodes/views.lua
@@ -1,7 +1,7 @@
-local nodes = require("kubectl.nodes")
-local commands = require("kubectl.commands")
-local tables = require("kubectl.view.tables")
-local actions = require("kubectl.actions")
+local actions = require("kubectl.actions.actions")
+local commands = require("kubectl.actions.commands")
+local nodes = require("kubectl.views.nodes")
+local tables = require("kubectl.utils.tables")
 
 local M = {}
