+++ lua/kubectl/views/secrets/views.lua
@@ -1,8 +1,8 @@
-local actions = require("kubectl.actions")
-local commands = require("kubectl.commands")
+local actions = require("kubectl.actions.actions")
+local commands = require("kubectl.actions.commands")
 local find = require("kubectl.utils.find")
-local secrets = require("kubectl.secrets")
-local tables = require("kubectl.view.tables")
+local secrets = require("kubectl.views.secrets")
+local tables = require("kubectl.utils.tables")
 
 local M = {}
